Undo #307; restore array-based Matrix constructors
PR #307 was incorrect, the array-based constructors are, in fact,
necessary.  Although constructing a matrix from a 2D-array variable
invokes the interop constructor:

    const float a[2][2] = {{1,0},{0,1}};
    M22f m(a);

constructing a matrix from a 2D array *parameter* uses the array
constructor, which #307 thought was never used, because it was never
invoked by the test suite:

    void foo (const float a[2][2])
    {
        M22f m(a);
    }

This restores the constructors and adds a test.

Signed-off-by: Cary Phillips <cary@ilm.com>