Replace =delete with SFINAE for invalid Vec integer methods
The Vec2, Vec3, and Vec4 classes declare methods `length()`,
`normalize()`, `normalizeExc()`, `normalizeNonNull()`, `normalized()`,
`normalizedExc()`, and `normalizedNonNull()`, which are invalid for
integer types.

Previously, these functions are declared for all template types but
the declarations are remove the integer instantiations via `=delete`:

  template <>
  IMATH_HOSTDEVICE short Vec2<short>::length () const IMATH_NOEXCEPT = delete;

However, this does not work with Clang20 on Windows msys2, as reported
in
https://github.com/AcademySoftwareFoundation/openexr/issues/2101. The
template appears to get instantiated _before_ the `=delete`, which
causes an error. Moving the `=delete` to _just_ after the class
declaration doesn't resolve the problem.

This change replaces the `=delete` with "Substitution Failure is Not
an Error" (SFINAE):

    template <typename U=T, typename std::enable_if<std::is_floating_point<U>::value, int>::type = 0>
    IMATH_HOSTDEVICE T length () const IMATH_NOEXCEPT;

With this form, the declaration is excluded entirely if the condition
fails (i.e. if T is not a floating-point type). Unlike `=delete`, which
makes the function ill-formed if called, SFINAE makes the function
invisible to the compiler when the condition is not satisfied.

This also adds CI jobs for C++20, C++23, C++26, and msys2 clang 20
(which lead to the original OpenEXR issue).

Signed-off-by: Cary Phillips <cary@ilm.com>